chore(rust): clippy/style cleanups across claude, settings, preflight
- Flatten claude/tests.rs to avoid module_inception; drop redundant printf arg
- Move ensure_gitignore above attachments test module (items_after_test_module)
- Settings tests: struct update syntax; Option::is_none_or for apiKeys assert
- errors tests: io::Error::other; extraction push ellipsis char; secret_present
  match empty str; database_migrations icon index without cast

+++ src-tauri/tests/database_migrations.rs
@@ -187,7 +187,7 @@ fn get_next_icon_index_cycles() {
             "connected",
             &format!("2026-01-{:02}T00:00:00Z", i + 1),
             None,
-            (i % 10) as i32,
+            i % 10,
         )
         .unwrap();
     }
